+++ assets/src/components/activities/common/Evaluation.tsx
@@ -23,11 +23,11 @@ export const Evaluation = ({ attemptState, context }:
   }
 
   return (
-    <div className={`evaluation feedback ${resultClass} my-1`}>
+    <div aria-label="result" className={`evaluation feedback ${resultClass} my-1`}>
       <div className="result">
-        <span className="score">{score}</span>
+        <span aria-label="score" className="score">{score}</span>
         <span className="result-divider">/</span>
-        <span className="out-of">{outOf}</span>
+        <span aria-label="out of" className="out-of">{outOf}</span>
       </div>
       <HtmlContentModelRenderer text={error ? errorText : feedback} context={context} />
     </div>